+++ utils.py
@@ -14,16 +14,30 @@ from kytos.core import log
 from kytos.core.events import KytosEvent
 
 
-def emit_event(controller, name, context="kytos/mef_eline", **kwargs):
+def map_evc_event_content(evc, **kwargs):
+    """Returns a set of values from evc to be used for content"""
+    return kwargs | {"evc_id": evc.id,
+                     "name": evc.name,
+                     "metadata": evc.metadata,
+                     "active": evc._active,
+                     "enabled": evc._enabled,
+                     "uni_a": evc.uni_a.as_dict(),
+                     "uni_z": evc.uni_z.as_dict()}
+
+
+def emit_event(controller, name, context="kytos/mef_eline", content=None):
     """Send an event when something happens with an EVC."""
     event_name = f"{context}.{name}"
-    event = KytosEvent(name=event_name, content=kwargs)
+    event = KytosEvent(name=event_name, content=content)
     controller.buffers.app.put(event)
 
 
 def notify_link_available_tags(controller, link, src_func=None):
     """Notify link available tags."""
-    emit_event(controller, "link_available_tags", link=link, src_func=src_func)
+    emit_event(controller, "link_available_tags", content={
+        "link": link,
+        "src_func": src_func
+    })
 
 
 def compare_endpoint_trace(endpoint, vlan, trace):